// src/app/main/pages/compartidos/interfaces/Itransaccion.ts
export interface Itransaccion {
    codigo?: number;
    codCliente?: number;
    codPersona?: number;
    codModulo?: number;
    codOperacion?: number;
    descripcion?: string;
    claveCuenta?: string;
    clave?: string;
    precio?: number;
    monto?: number;
    numProducto?: number;
    numProductoAnt?: number;
    numExistenciaActual?: number;
    numMes?: number;
    numDiasExtra?: number;
    fechaInicio?: string;
    fechaFin?: string;
    fechaRegistra?: string;
    fechaCambia?: string;
    estado?: string;

    cliente?: any,
    producto?: any,
    modulo?: any,
    operacion?: any,
    colorFila?: string;
    visibleBoton?: string;
    colorColumna?: string;
    descripcionProducto?: string;
    precioCosto?: number;
    precioMayoreo?: number;
    nombreCliente?: string;
    prefijoTelefonico?: string;
    celular?: string;
    numDiasRenovar: number;
    displayNoneListaCuentaClave: string;
    grupo?: string;
    medida?: string;
}